+++ Core/Runtime/ResKit/Loader/UniTask/DefaultResLoaderUniTask.cs
@@ -1,5 +1,4 @@
 #if YOKIFRAME_UNITASK_SUPPORT
-using System;
 using System.Threading;
 using Cysharp.Threading.Tasks;
 using UnityEngine;
@@ -16,32 +15,11 @@ namespace YokiFrame
     }
 
     /// <summary>
-    /// 默认 UniTask 加载器（Resources）
+    /// 默认 UniTask 加载器（Resources） - 继承 DefaultResLoader，仅扩展 UniTask 异步方法
     /// </summary>
-    public class DefaultResLoaderUniTask : IResLoaderUniTask
+    public class DefaultResLoaderUniTask : DefaultResLoader, IResLoaderUniTask
     {
-        private readonly IResLoaderPool mPool;
-        private Object mAsset;
-
-        public DefaultResLoaderUniTask(IResLoaderPool pool) => mPool = pool;
-
-        public T Load<T>(string path) where T : Object
-        {
-            mAsset = Resources.Load<T>(path);
-            ResLoadTracker.OnLoad(this, path, typeof(T), mAsset);
-            return mAsset as T;
-        }
-
-        public void LoadAsync<T>(string path, Action<T> onComplete) where T : Object
-        {
-            var request = Resources.LoadAsync<T>(path);
-            request.completed += _ =>
-            {
-                mAsset = request.asset;
-                ResLoadTracker.OnLoad(this, path, typeof(T), mAsset);
-                onComplete?.Invoke(mAsset as T);
-            };
-        }
+        public DefaultResLoaderUniTask(IResLoaderPool pool) : base(pool) { }
 
         public async UniTask<T> LoadUniTaskAsync<T>(string path, CancellationToken cancellationToken = default) where T : Object
         {
@@ -50,21 +28,6 @@ namespace YokiFrame
             mAsset = request.asset;
             ResLoadTracker.OnLoad(this, path, typeof(T), mAsset);
             return mAsset as T;
-        }
-
-        public void UnloadAndRecycle()
-        {
-            ResLoadTracker.OnUnload(this);
-            if (mAsset != default)
-            {
-                // GameObject/Component 无法通过 UnloadAsset 释放，置空引用等待 UnloadUnusedAssets
-                if (mAsset is not (GameObject or Component))
-                {
-                    Resources.UnloadAsset(mAsset);
-                }
-                mAsset = null;
-            }
-            mPool.Recycle(this);
         }
     }
 }